Clean up typescript errors etc
- tweak configuration details
- reorder code
